strona_heroku/strona_heroku/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colored==1.4.2
colorlog==4.1.0
colormap==1.0.2
constantly==15.1.0
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
django-heroku==0.3.1
easydev==0.9.38
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
pexpect==4.8.0
Pillow==7.0.0
psycopg2==2.8.4
ptyprocess==0.6.0
PyDispatcher==2.0.5
pygame==1.9.6
PyHamcrest==2.0.1
pylint==2.4.4
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.13.0
smmap2==2.0.5
soupsieve==2.0
sqlparse==0.3.0
stevedore==1.31.0
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.7.1
